Adds milliseconds capability to start time
Users can now enter milliseconds for the start times. This will help
with timing on some tracks. Fixes a bug in the migration mechanism as
well.

+++ power_hour_creator/db/migrations/1.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS tracks (
   url         text NOT NULL,
   title       text NOT NULL,
   length      integer NOT NULL,
-  start_time  integer NOT NULL,
+  start_time  decimal(7, 3) NOT NULL,
   full_song   integer NOT NULL DEFAULT 0,
   power_hour_id integer NOT NULL, FOREIGN KEY(power_hour_id) REFERENCES power_hours(id)
 );
